修复编译错误和类型转换问题
- 修复StudentMockExamService.cs中变量名大小写不一致问题（totalScoreFromRules）
- 修复double到int的类型转换错误（MockExamTotalScore）
- 修复double到decimal的类型转换错误（TotalScore、Score等属性）
- 添加必要的类型转换，确保数据类型兼容性
- 解决由于ModuleDto.Score类型从int改为double导致的连锁反应

所有编译错误已修复，项目可以正常编译。

+++ ExamLab/Models/ExamModule.cs
@@ -168,7 +168,7 @@ public class ExamModule : ReactiveObject
     /// <summary>
     /// 模块分值
     /// </summary>
-    [Reactive] public int Score { get; set; }
+    [Reactive] public double Score { get; set; }
 
     /// <summary>
     /// 模块包含的题目
